+++ apps/web/src/entities/feed/index.ts
@@ -1,5 +1,5 @@
 export { formatLikeCount, formatStudyDuration } from './lib/format';
-export { FEED_QUERY_KEY } from './model/queryKeys';
+export { FEED_QUERY_KEY } from './model/query-keys';
 export type { FeedPage, FeedPost, FeedTag } from './model/types';
 export { default as FeedStatsSummary } from './ui/FeedStatsSummary';
 export { default as TagBadgeGroup } from './ui/TagBadgeGroup';
